♻️ refactor [#24] : hub, hubroute 모듈 분리 및 dto 리팩토리
- hub, hubroute 모듈 분리
- dto 를 request, response 디렉토리로 분리
- 불필요한 deleteDto 삭제
