--- com.sparta.logistics.hub-service/src/main/java/com/sparta/logistics/hub_service/application/dto/HubDeleteResponseDto.java
@@ -1,13 +1,0 @@
-package com.sparta.logistics.hub_service.application.dto;
-
-import java.math.BigDecimal;
-
-public class HubDeleteResponseDto {
-
-  private long userId;
-  private String hubName;
-  private String address;
-  private BigDecimal latitude;
-  private BigDecimal longitude;
-
-}
