--- com.sparta.logistics.hub-service/src/main/java/com/sparta/logistics/hub_service/application/dto/HubRouteListResponseDto.java
@@ -1,15 +1,0 @@
-package com.sparta.logistics.hub_service.application.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@Builder
-public class HubRouteListResponseDto {
-  private String startHubName;
-  private String endHubName;
-  private Long deliveryTime;
-  private Long deliveryDistance;
-}
