+++ com.sparta.logistics.hub-service/src/main/java/com/sparta/logistics/hub_service/hub/application/dto/request/HubCreateRequestDto.java
@@ -1,18 +1,18 @@
-package com.sparta.logistics.hub_service.application.dto;
+package com.sparta.logistics.hub_service.hub.application.dto.request;
 
 import java.math.BigDecimal;
 import lombok.Builder;
 import lombok.Getter;
+import lombok.Setter;
 
-@Builder
 @Getter
-public class HubUpdateResponseDto {
+@Setter
+@Builder
+public class HubCreateRequestDto {
+
   private long userId;
   private String hubName;
   private String address;
   private BigDecimal latitude;
   private BigDecimal longitude;
-
-  // 테스트를 위한 임시 추가
-  private long hubId;
 }
