+++ com.sparta.logistics.hub-service/src/main/java/com/sparta/logistics/hub_service/hub/application/dto/request/HubUpdateRequestDto.java
@@ -1,7 +1,6 @@
-package com.sparta.logistics.hub_service.application.dto;
+package com.sparta.logistics.hub_service.hub.application.dto.request;
 
 import java.math.BigDecimal;
-import java.util.UUID;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
@@ -9,16 +8,12 @@ import lombok.Setter;
 @Getter
 @Setter
 @Builder
-public class HubDetailResponseDto {
-
+public class HubUpdateRequestDto {
   private long userId;
   private String hubName;
   private String address;
   private BigDecimal latitude;
   private BigDecimal longitude;
-
-  // 테스트를 위한 임시 추가
-  private long hubId;
 
 
 }
