+++ com.sparta.logistics.hub-service/src/main/java/com/sparta/logistics/hub_service/hub/application/dto/response/HubDetailResponseDto.java
@@ -1,4 +1,4 @@
-package com.sparta.logistics.hub_service.application.dto;
+package com.sparta.logistics.hub_service.hub.application.dto.response;
 
 import java.math.BigDecimal;
 import lombok.Builder;
@@ -8,11 +8,16 @@ import lombok.Setter;
 @Getter
 @Setter
 @Builder
-public class HubCreateRequestDto {
+public class HubDetailResponseDto {
 
   private long userId;
   private String hubName;
   private String address;
   private BigDecimal latitude;
   private BigDecimal longitude;
+
+  // 테스트를 위한 임시 추가
+  private long hubId;
+
+
 }
