+++ com.sparta.logistics.hub-service/src/main/java/com/sparta/logistics/hub_service/hubroute/application/dto/request/HubRouteCreateRequestDto.java
@@ -1,4 +1,4 @@
-package com.sparta.logistics.hub_service.application.dto;
+package com.sparta.logistics.hub_service.hubroute.application.dto.request;
 
 import java.util.UUID;
 import lombok.Builder;
