+++ com.sparta.logistics.hub-service/src/main/java/com/sparta/logistics/hub_service/hubroute/application/dto/response/HubRouteCreateResponseDto.java
@@ -1,8 +1,13 @@
-package com.sparta.logistics.hub_service.application.dto;
+package com.sparta.logistics.hub_service.hubroute.application.dto.response;
 
 import java.util.UUID;
+import lombok.Builder;
+import lombok.Getter;
 
-public class HubRouteDeleteResponseDto {
+@Getter
+@Builder
+public class HubRouteCreateResponseDto {
+
   private UUID startHubId;
   private UUID endHubId;
   private String startHubName;
