+++ com.sparta.logistics.hub-service/src/main/java/com/sparta/logistics/hub_service/hubroute/application/dto/response/HubRouteDetailResponseDto.java
@@ -1,12 +1,14 @@
-package com.sparta.logistics.hub_service.application.dto;
+package com.sparta.logistics.hub_service.hubroute.application.dto.response;
 
 import java.util.UUID;
 import lombok.Builder;
 import lombok.Getter;
+import lombok.Setter;
 
 @Getter
 @Builder
-public class HubRouteCreateResponseDto {
+@Setter
+public class HubRouteDetailResponseDto {
 
   private UUID startHubId;
   private UUID endHubId;
@@ -14,4 +16,7 @@ public class HubRouteCreateResponseDto {
   private String endHubName;
   private Long deliveryTime;
   private Long deliveryDistance;
+
+  private Long id; // 임시
+
 }
