+++ com.sparta.logistics.hub-service/src/main/java/com/sparta/logistics/hub_service/hubroute/application/dto/response/HubRouteListResponseDto.java
@@ -1,14 +1,13 @@
-package com.sparta.logistics.hub_service.application.dto;
+package com.sparta.logistics.hub_service.hubroute.application.dto.response;
 
-import java.util.UUID;
 import lombok.Builder;
 import lombok.Getter;
+import lombok.Setter;
 
 @Getter
+@Setter
 @Builder
-public class HubRouteUpdateResponseDto {
-  private UUID startHubId;
-  private UUID endHubId;
+public class HubRouteListResponseDto {
   private String startHubName;
   private String endHubName;
   private Long deliveryTime;
