+++ com.sparta.logistics.hub-service/src/main/java/com/sparta/logistics/hub_service/hubroute/data/HubRouteData.java
@@ -1,7 +1,7 @@
-package com.sparta.logistics.hub_service.data;
+package com.sparta.logistics.hub_service.hubroute.data;
 
-import com.sparta.logistics.hub_service.application.dto.HubRouteDetailResponseDto;
-import com.sparta.logistics.hub_service.application.dto.HubRouteListResponseDto;
+import com.sparta.logistics.hub_service.hubroute.application.dto.response.HubRouteDetailResponseDto;
+import com.sparta.logistics.hub_service.hubroute.application.dto.response.HubRouteListResponseDto;
 import java.util.List;
 import java.util.UUID;
